demo/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
elasticsearch-dsl==7.4.1
PyPDF2==3.0.1
Pillow==10.1.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-filter==23.4
setuptools